C API and GUI: Handle the case when the number of bins can't be determined
If a font is not load yet the size of the graph's label can have some
nonsensical value, hence the number of bins can be nonsensical as well.
In this case set only the range of the model. The rest of the work will
be done when the font is loaded.

Signed-off-by: Yordan Karadzhov (VMware) <y.karadz@gmail.com>