GUI: Connect the Color scheme slider and the Table view widget
When the Color scheme of the Graph widget changes the same must happen in
the table. Note that this has an effect only if more than one data files
are loaded. Otherwise  the Table view widget shows no colors.

When the Color scheme of the Graph widget changes the same must happen in the table.

Signed-off-by: Yordan Karadzhov (VMware) <y.karadz@gmail.com>