GUI: Block the "Add Time Offset" menu if it is inapplicable.
The menu is used to add time offset to one of the data streams. Offset
can be added only to the appended streams, while the prior stream always
keeps its original time-stamp values. Because of this we want to enable
the "Add Time Offset" menu only in the case of multiple data streams.

Signed-off-by: Yordan Karadzhov (VMware) <y.karadz@gmail.com>