C API: Allow negative values for the range edges of the model
We don't want the model to be extended to negative values and this
cannot realy happen indeed. However, in some corner cases when we
calculate the range we may get nevative values. When those negative
values are casted to unsigned types the whole logic that processes
the state of the model gets corrupted, resulting into segmentation
faults.

Signed-off-by: Yordan Karadzhov (VMware) <y.karadz@gmail.com>