GUI: Make the "stop search" button always visible
So far the button was showing up only when we do a single-threaded
search in the data. Stopping the multi-threaded searches became
possible after the modification of the parallel search. When the
multi-threaded search is restarted (after being stopped) the label
showing the number of matches found is reset to show nothing. It
will show the count again when the search finishes or is stopped.

Signed-off-by: Yordan Karadzhov (VMware) <y.karadz@gmail.com>