GUI: Fix bug in KsDataStore::_openDataFile()
When we open a new file, the GUI will first close all existing data
streams. Hence the Id of the new stream returned by kshark_open()
must always be 0 (the Id of the very first stream). However when
we append file, the already opened streams are not closed, so the
returned stream Id can not be 0.

Signed-off-by: Yordan Karadzhov (VMware) <y.karadz@gmail.com>